Updated project structure
Added the following:
- cmd & test folders, main.go and tourplan-app.yaml file moved to cmd file.
- Makefile with lint, formatting and build only.
- README file
- go.mod file
