+++ .gcloudignore
@@ -23,3 +23,6 @@
 *.test
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+
+# OS generated files 
+*.DS_Store
