+++ .gitignore
@@ -17,5 +17,8 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
+# OS generated files 
+*.DS_Store
+
 # Go workspace file
 go.work
