+++ README.md
@@ -1,0 +1,3 @@
+#Tourist Planner Server App 
+
+This project describes the server application for the TourPlannit iOS app in https://github.com/SvetoslavAngelov/TourPlannit.
