--- app.yaml
@@ -1,5 +1,0 @@
-runtime: go119
-instance_class: F1
-
-automatic_scaling: 
-  max_instances: 1
