+++ cmd/tourplan-app.yaml
@@ -1,0 +1,5 @@
+runtime: go119
+instance_class: F1
+
+automatic_scaling: 
+  max_instances: 1
