+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/SvetoslavAngelov/tourplan-app
+
+go 1.19
